Move _get_size from generator to cupy.__init__ and use it in ndarray
